Create README.md
Currently this API allows for simple submission of name, email and message to a database. Looking to add logic that when submitted will take data and email it to intended email of choice. Connect your database to the config folder/default.json and test with a post request in Postman.